% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conda_create_silentJSON}
\alias{conda_create_silentJSON}
\title{Silent/json version of reticulate's conda_create}
\usage{
conda_create_silentJSON(
  envname = NULL,
  forge = TRUE,
  channel = character(),
  conda = "auto"
)
}
\arguments{
\item{forge}{Boolean; include the [Conda Forge](https://conda-forge.org/)
repository?}

\item{channel}{An optional character vector of Conda channels to include.
When specified, the `forge` argument is ignored. If you need to
specify multiple channels, including the Conda Forge, you can use
`c("conda-forge", <other channels>)`.}
}
\value{
Nothing returned, conda environment created.
}
\description{
Reticulate's conda_create with silent output and json output capability
}
\keyword{internal}
