% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conda_install_silentJSON}
\alias{conda_install_silentJSON}
\title{Silent/json version of reticulate's conda_install}
\usage{
conda_install_silentJSON(
  envname = NULL,
  packages,
  forge = TRUE,
  channel = character(),
  pip = FALSE,
  pip_options = character(),
  pip_ignore_installed = FALSE,
  conda = "auto",
  python_version = NULL,
  ...
)
}
\arguments{
\item{forge}{Boolean; include the [Conda Forge](https://conda-forge.org/)
repository?}

\item{channel}{An optional character vector of Conda channels to include.
When specified, the `forge` argument is ignored. If you need to
specify multiple channels, including the Conda Forge, you can use
`c("conda-forge", <other channels>)`.}

\item{pip_options}{An optional character vector of additional command line
arguments to be passed to `pip` if `pip` is used.}

\item{pip_ignore_installed}{Ignore installed versions when using pip. This is
`TRUE` by default so that specific package versions can be installed even
if they are downgrades. The `FALSE` option is useful for situations where
you don't want a pip install to attempt an overwrite of a conda binary
package (e.g. SciPy on Windows which is very difficult to install via pip
due to compilation requirements).}
}
\description{
Reticulate's conda_install with silent output and json output capability
}
\keyword{internal}
