% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{export_CondaEnv}
\alias{export_CondaEnv}
\title{Export Conda environment.}
\usage{
export_CondaEnv(
  env_name,
  yml_export = NULL,
  pathToMiniConda = NULL,
  depends = TRUE
)
}
\arguments{
\item{env_name}{Name of environment you want to save}

\item{yml_export}{Destination for exported environment yml file}

\item{pathToMiniConda}{NULL Path to miniconda installation}

\item{depends}{if FALSE will only include packages explicitly installed and not dependencies}
}
\value{
Nothing returned. Output written to file.
}
\description{
Export Conda environment
}
\author{
Matt Paul
}
