% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_CondaEnv}
\alias{import_CondaEnv}
\title{Import Conda environment.}
\usage{
import_CondaEnv(yml_import, name = NULL, pathToMiniConda = NULL)
}
\arguments{
\item{yml_import}{conda environment yml file}

\item{pathToMiniConda}{NULL Path to miniconda installation}
}
\value{
Nothing returned. Output written to file.
}
\description{
Import Conda environment
}
\author{
Matt Paul
}
