% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_CondaPkgs.R
\name{list_CondaPkgs}
\alias{list_CondaPkgs}
\title{List Conda packages.}
\usage{
list_CondaPkgs(env, pathToMiniConda = NULL, pkg = NULL)
}
\arguments{
\item{env}{environment to look in}

\item{pathToMiniConda}{NULL Path to miniconda installation}

\item{pkg}{Package name. If this is supplied to list_CondaPkg, it will query whether that package is present in the given environment.}
}
\value{
Conda package information is printed to the screen. If package name is supplied a TRUE/FALSE will be returned depending on whether that package is present or not.
}
\description{
List Conda packages
}
\details{
This function will list all packages from a specific conda environment.
For a more detailed walkthrough check our webpage: https://rockefelleruniversity.github.io/Herper_Page/
}
\author{
Matt Paul
}
