% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conda_install_silentJSON}
\alias{conda_install_silentJSON}
\title{Silent/json version of reticulate's conda_install}
\usage{
conda_install_silentJSON(
  envname = NULL,
  packages,
  forge = TRUE,
  channel = character(),
  conda = "auto",
  ...
)
}
\arguments{
\item{forge}{Boolean; include the [Conda Forge](https://conda-forge.org/)
repository?}

\item{channel}{An optional character vector of Conda channels to include.
When specified, the `forge` argument is ignored. If you need to
specify multiple channels, including the Conda Forge, you can use
`c("conda-forge", <other channels>)`.}
}
\description{
Reticulate's conda_install with silent output and json output capability
}
\keyword{internal}
