% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{install_CondaSysReqs}
\alias{install_CondaSysReqs}
\title{Install Conda requirements listed in the System Requirement field of description}
\usage{
install_CondaSysReqs(
  pkg,
  channels = NULL,
  env = NULL,
  pathToMiniConda = NULL,
  updateEnv = FALSE,
  SysReqsAsJSON = TRUE
)
}
\arguments{
\item{pkg}{Package to install Conda System Requirements from.}

\item{channels}{Additional channels for miniconda (bioconda defaults and conda-forge are included automatically)}

\item{env}{Name of Conda environment to install tools into.}

\item{pathToMiniConda}{NULL Path to miniconda installation}

\item{updateEnv}{Update existing package's conda environment if already installed.}

\item{SysReqsAsJSON}{Parse the SystemRequirements in JSON format (see Details). Default is TRUE.}
}
\value{
Nothing returned. Output written to file.
}
\description{
Install Conda requirements
}
\examples{
testPkg <- system.file("extdata/HerperTestPkg",package="CondaSysReqs")
install.packages(testPkg,type = "source",repos = NULL)
condaPaths <- install_CondaSysReqs("HerperTestPkg",pathToMiniConda=tempdir(),SysReqsAsJSON=FALSE)
system2(file.path(condaPaths$pathToEnvBin,"samtools"),args = "--help")
}
\author{
Thomas Carroll
}
