% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{install_CondaTools}
\alias{install_CondaTools}
\title{Install Conda requirements.}
\usage{
install_CondaTools(
  tools,
  env,
  channels = NULL,
  pathToMiniConda = NULL,
  updateEnv = FALSE
)
}
\arguments{
\item{tools}{software to install using conda.}

\item{env}{Name of Conda environment to install tools into.}

\item{channels}{Additional channels for miniconda (bioconda defaults and conda-forge are included automatically)}

\item{pathToMiniConda}{NULL Path to miniconda installation}

\item{updateEnv}{Update existing package's conda environment if already installed.}
}
\value{
Nothing returned. Output written to file.
}
\description{
Install Conda requirements
}
\examples{
condaPaths <- install_CondaTools("salmon","salmon",pathToMiniConda=tempdir())
system2(file.path(condaPaths$pathToEnvBin,"salmon"),args = "--help")
}
\author{
Thomas Carroll
}
