% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{install_CondaTools}
\alias{install_CondaTools}
\title{Install Conda requirements.}
\usage{
install_CondaTools(
  tools,
  env,
  channels = NULL,
  pathToMiniConda = NULL,
  updateEnv = FALSE,
  search = TRUE
)
}
\arguments{
\item{tools}{Vector of software to install using conda.}

\item{env}{Name of Conda environment to install tools into.}

\item{channels}{Additional channels for miniconda (bioconda defaults and conda-forge are included automatically)}

\item{pathToMiniConda}{NULL Path to miniconda installation}

\item{updateEnv}{Update existing package's conda environment if already installed.}

\item{search}{Whether to search for the package name and version before installing. It is highly recommended this be set to TRUE as information about available versions or similar packages will be included in the output if the exact match is not found.}
}
\value{
Nothing returned. Output written to file.
}
\description{
Install Conda requirements
}
\examples{
condaDir <- file.path(tempdir(), "r-miniconda")
condaPaths <- install_CondaTools("salmon", "salmon", pathToMiniConda = condaDir)
system2(file.path(condaPaths$pathToEnvBin, "salmon"), args = "--help")
}
\author{
Thomas Carroll
}
