% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_CondaEnv.R
\name{list_CondaEnv}
\alias{list_CondaEnv}
\title{List Conda environments.}
\usage{
list_CondaEnv(pathToMiniConda = NULL, allCondas = FALSE, env = NULL)
}
\arguments{
\item{pathToMiniConda}{NULL Path to miniconda installation}

\item{allCondas}{Logical. Whether to return conda environments, for all discoverable conda installs, or just the conda specified in pathToMiniConda.}

\item{env}{Environment name. If this is supplied to list_CondaEnv, it will query whether that environment is present in the given conda.}
}
\value{
Conda environment names and the file paths to their conda installation are printed to the screen. If environment name is supplied a TRUE/FALSE will be returned depending on whether that environment is present or not.
}
\description{
List Conda environments
}
\examples{
condaPaths <- install_CondaTools("salmon", "herper_env")
list_CondaEnv()
list_CondaEnv( env = "herper_env")
}
\author{
Matt Paul
}
