% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{export_CondaEnv}
\alias{export_CondaEnv}
\alias{import_CondaEnv}
\title{Export Conda environment.}
\usage{
export_CondaEnv(
  env_name,
  yml_export = NULL,
  pathToMiniConda = NULL,
  depends = TRUE
)

import_CondaEnv(
  yml_import,
  name = NULL,
  pathToMiniConda = NULL,
  install = TRUE,
  channels = NULL
)
}
\arguments{
\item{env_name}{Name of environment you want to save}

\item{yml_export}{Destination for exported environment yml file}

\item{pathToMiniConda}{NULL Path to miniconda installation}

\item{depends}{if FALSE will only include packages explicitly installed and not dependencies}

\item{yml_import}{conda environment yml file}

\item{name}{Name of the environment to create.}

\item{install}{TRUE/FALSE whether to install miniconda at path if it doesn't exist.}

\item{channels}{Channels for miniconda (bioconda and conda-forge are defaults).}
}
\value{
Nothing returned. Output written to file.

Nothing returned. Output written to file.
}
\description{
Export Conda environment

Import Conda environment
}
\author{
Matt Paul
}
