% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_search.R
\name{conda_search}
\alias{conda_search}
\title{Search package availabilty}
\usage{
conda_search(package, channel = NULL, print_out = TRUE, pathToMiniConda = NULL)
}
\arguments{
\item{package}{Package to search for. If an exact match is found, the funtion will return true (assuming 'package_version' is left NULL or is a valid entry). If there is not an exact match and other packages contain this text, the function will return FALSE but the alternative options will be printed if print_out = TRUE.}

\item{channel}{A specific to search in addition to defaults (bioconda defaults and conda-forge are the default channels)}

\item{print_out}{Either True or FALSE indicating whether to print out information about available builds and channels for the search entry.}

\item{pathToMiniConda}{Path to miniconda installation. If this is set to NULL (default), then the output of 'reticulate::miniconda_path()' is used.}
}
\value{
TRUE/FALSE
}
\description{
Search package availabilty
}
\examples{
condaDir <- file.path(tempdir(), "r-miniconda")
if(!dir.exists(condaDir)){
 condaPaths <- install_CondaTools("salmon", "salmon", pathToMiniConda = condaDir)
}
conda_search("salmon",pathToMiniConda=condaDir)
}
\author{
Doug Barrows and Matt Paul
}
