% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCondaPaths.R
\name{with_CondaEnv}
\alias{with_CondaEnv}
\alias{local_CondaEnv}
\title{Use Conda environments.}
\usage{
with_CondaEnv(
  new,
  code,
  actionPATH = "prefix",
  actionPYTHONLIB = "replace",
  actionPERL5LIB = "replace",
  actionJAVALIB = "replace"
)

local_CondaEnv(
  new,
  actionPATH = "prefix",
  actionPYTHONLIB = "replace",
  actionPERL5LIB = "replace",
  actionJAVALIB = "replace",
  .local_envir = parent.frame()
)
}
\arguments{
\item{env}{conda environment}

\item{pathToMiniConda}{Path to miniconda.}
}
\description{
Use Conda environments

Use Conda environments
}
\author{
Thomas Carroll

Thomas Carroll
}
