% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_conda.R
\name{with_CondaEnv}
\alias{with_CondaEnv}
\alias{local_CondaEnv}
\title{Use Conda environments.}
\usage{
with_CondaEnv(
  new,
  code,
  pathToMiniConda = NULL,
  path_action = "prefix",
  path_additional = NULL,
  pythonpath_action = "replace",
  perl5lib_action = "replace",
  javahome_action = "replace",
  pythonpath_additional = NULL,
  perl5lib_additional = NULL,
  javahome_additional = NULL
)

local_CondaEnv(
  new,
  pathToMiniConda = NULL,
  path_action = "prefix",
  path_additional = NULL,
  pythonpath_action = "replace",
  perl5lib_action = "replace",
  javahome_action = "replace",
  pythonpath_additional = NULL,
  perl5lib_additional = NULL,
  javahome_additional = NULL,
  .local_envir = parent.frame()
)
}
\arguments{
\item{new}{The name of conda environment to include in the temporary R environment.}

\item{code}{Code to execute in the temporary R environment}

\item{pathToMiniConda}{Path to miniconda.}

\item{path_action}{Should new values "replace", "prefix" or "suffix" existing PATH variable.}

\item{path_additional}{Additional paths to suffix to existing PATH variable.}

\item{pythonpath_action}{Should new values "replace", "prefix" or "suffix" existing PYTHONPATH variable.}

\item{perl5lib_action}{Should new values "replace", "prefix" or "suffix" existing PERL5LIB variable.}

\item{javahome_action}{Should new values "replace", "prefix" or "suffix" existing JAVA_HOME variable.}

\item{pythonpath_additional}{Additional paths to suffix to existing PYTHONPATH variable.}

\item{perl5lib_additional}{Additional paths to suffix to existing PERL5LIB variable.}

\item{javahome_additional}{Additional paths to suffix to existing JAVA_HOME variable.}

\item{.local_envir}{The environment to use for scoping.}
}
\value{
Nothing returned.
}
\description{
Use Conda environments

Use Conda environments
}
\examples{
testYML <- system.file("extdata/HerperTestPkg_0.1.0.yml",package="CondaSysReqs")
condaDir <- file.path(tempdir(),"r-miniconda")
import_CondaEnv(testYML,"HerperTest",pathToMiniConda=condaDir)
with_CondaEnv("HerperTest",system2(command = "rmats.py",args = "-h"),pathToMiniConda = condaDir)
\dontrun{
  install_CondaTools("cytoscape","cytoscape",updateEnv = TRUE,pathToMiniConda = condaDir)
  with_CondaEnv("cytoscape",system2(command = "cytoscape.sh"),pathToMiniConda = condaDir)
}
}
\author{
Thomas Carroll
}
